pyproject.toml
[project]
name = "hello-mcp-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-anthropic>=0.3.15",
    "mcp-use>=1.3.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
